requirements.txt
rembg
Pillow
onnxruntime
torch
transformers
diffusers
accelerate
fastapi
uvicorn[standard]
python-multipart